+++ CHANGELOG.md
@@ -1,3 +1,82 @@
+### v0.18.1.pre
+
+* bug-fixes
+  * Prevent logging `code` and `refresh_token` on `/api/token` calls
+
+### v0.18.0.pre
+
+* features
+  * Add new token retrieval/refresh methods
+    * `Token#authorization_code`
+    * `Token#refresh_token`
+    * `Token#client_credentials`
+  * Add alias methods for avoiding deprecation
+    * `refresh` -> `refresh_token`
+    * `create`  -> `authorization_code`
+
+### v0.17.0.pre
+
+* enhancements
+  * Add customizable pagination size when sending `paginated: false` parameter to `#all` requests.
+
+### v0.16.1.pre
+
+* bug-fixes
+  * Fix method to retrieve resources without pagination
+
+### v0.16.0.pre
+
+* features
+  * Add pickup options methods
+    * `PickupOptions#delete`
+    * `PickupOptions#find`
+    * `PickupOptions#sort`
+    * `PickupOptions#update`
+
+### v0.15.0.pre
+
+* features
+  * Add pickup options methods
+    * `PickupOptions#all`
+    * `PickupOptions#create`
+
+### v0.14.0.pre
+
+* deprecations
+  * Remove `Products#assign_variation_images_differentiator`
+
+* features
+  * Add `Products#assign_variation_attribute_as_differentiator`
+
+### v0.13.0.pre
+
+* features
+  * Add payment method definitions methods
+    * `PaymentMethodDefinitions#all`
+    * `PaymentMethodDefinitions#create`
+    * `PaymentMethodDefinitions#delete`
+    * `PaymentMethodDefinitions#find`
+    * `PaymentMethodDefinitions#update`
+  * Add possibility to log request `headers` and `bodies`
+
+### v0.12.1.pre
+
+* bug-fixes
+  * Fix camelize function for hashes containing arrays
+
+### v0.12.0.pre
+
+* features
+  * Add locations methods
+    * `Locations#all`
+    * `Locations#create`
+    * `Locations#delete`
+    * `Locations#find`
+    * `Locations#update`
+
+* enhancements
+  * Use `autoload` instead of `require`
+
 ### v0.11.1.pre
 
 * bug-fixes